made change so that issue updates generate files with unique names
in order to support one master issues database in the pits and the slave
in the cloud, made changes so that any issue update results in a new
issue file being created and sequenced such that updates are provided in
proper order. I still haven't dealt with issue comments just yet, but
that should be straightforward.