Updated the code generator to account for the project.properties file
The recent changes to the base scouting application project to reference
a separate library for the file sync utilities required a corresponding
change to the code generator application which creates a new project
from the base. Some new code was needed to update the library references
in the project.properties file.