Modified to support the new layout of the site.
Modified the main web server module and most of the WebXxx.py modules to
build only HTML fragments that can be rendered through the new layout
template.

Additional changes will be made to the how the pages are generated but
for now, these changes allow all pages to be supported within the new
layout. The new changes will generally involve the creation of a new
web.py templates, inclusion of more pure javascript code, and the
removal of code that currently generates the HTML to be rendered.