Fixed problem where match data was missing from team page
The match data table on the team data page wasn't getting filled in.
There is a filter that is defined in the attributes definition file that
must be defined in order for the match data to show up. Made changes to
key off the "Include_In_Report" setting if the "Include_In_Team_Display"
setting is missing.