Extracted the bluetooth sync server to its own module
Extracted the bluetooth sync server processing to a separate class now
that it will be set up to run as part of the web server application.
Also, updated the bluetooth sync behavior itself to match the changes
made to the Android side of things and to support the transfer of binary
files along with the existing support for text file transfer.