Added support to modify the attribute weights via the website
Previously, the only way to update the attribute weights would be to
modify the attribute definitions spreadsheet and then have the web
server read in the new definitions and recalculate the weighted scoring.
Now, there is a menu option that presents a form with all valid
attributes and their current weights. Any changes are captured in an
overrides file that is then used to recalculate the weighted scoring