Modified to support the updated datafeed mechanism and other changes
Modified the event and match request processing to pass additional
information to the functions that are preparing the event and match data
pages. Also added a URL support for setting the attribute weights. And,
made some minor changes to attempt to eliminate the SQLite exception
messages complaining about accessing connections from multiple threads.