Enhanced Base Scouting App to make setup easier
Made a number of changes to the base scouting app from which the other
apps are derived:

1. Added separate competition configuration file and menu option. Now,
the competition configuration can be set manually on each tablet, or
downloaded from the central server.

2. Added separate device control setting to allow for syncing of the
scouting data files, scouting media files or both. Syncing a lot of
pictures takes time and we wanted to control that part of the sync
separately.

3. Modified the discard button behavior to provide the option to save
the scouting data prior to the discard operation. Also relabeled the
button on the user interface to say Clear, instead of Discard.