Fix bug where the utility function returned unexpected event info
The get_event_info() utility function was returning an empty string list
when the event itself didn't exist. And, the HTML template didn't
properly handle this condition. The fix was to return nothing for this
condition by the utility function and for the template to handle the
call that returns nothing. Ultimately, the templates weren't currently
using the event information string anyway, so I just removed those calls
from the template.