Fixed error where exception thrown during db dump
Fixed a case where an exception was thrown due to an empty record when
dumping scouting database records. Problem was caused by a gap in the
column_order specified in the attributes definition file. Code was added
to handle that particular format error and gracefully continue without
throwing the exception.